+++ dad-is-great-in-C/s_sort.cpp
@@ -11,7 +11,7 @@ int main() {
   std::cin >> n;
   auto st = std::chrono::system_clock::now();
 
-  std::vector<unsigned int> a(n);
+  std::vector<unsigned long int> a(n);
   std::random_device rd;
   std::mt19937 gen(rd());
   std::uniform_int_distribution<std::mt19937::result_type> dist(1, 1000000);
